Fixed build warnings on windows
git-svn-id: https://svn.code.sf.net/p/veroroute/code/trunk@36 d5f823ff-b8c7-416f-ab5d-d69718ecf442

+++ src/CompManager.h
@@ -139,7 +139,7 @@ public:
 
 		const Component* pPrev(nullptr);	// The previous wire on the line
 		const Component* pLast(nullptr);	// The wire that reaches most along the line
-		size_t i(0);
+		int i(0);
 		for (auto& p : wiresH)
 		{
 			if ( pPrev == nullptr || p->GetRow() != pPrev->GetRow() )	// Reset all if new row
